refactor: extract inject decorator logic into a reusable function
